Pull in patch from current: Fix (mickey): -- on `out of data' condition print a warning and exit w/ whatever current checksum we have gotten accumulated. this is somewhat conformant w/ what other archs are doing. from Camiel Dobbelaar <cd@sentia.nl> -- fix preposition in the printf msg

+++ usr.sbin/rtadvd/advcap.h
@@ -1,3 +1,6 @@
+/*	$OpenBSD: advcap.h,v 1.3 2000/05/23 11:23:22 itojun Exp $	*/
+/*	$KAME: advcap.h,v 1.3 2001/02/01 09:12:08 jinmei Exp $	*/
+
 /*
  * Copyright (C) 1994,1995 by Andrey A. Chernov, Moscow, Russia.
  * All rights reserved.
@@ -35,7 +38,7 @@ __BEGIN_DECLS
 
 extern int agetent __P((char *, const char *));
 extern int agetflag __P((const char *));
-extern int agetnum __P((const char *));
+extern long long agetnum __P((const char *));
 extern char *agetstr __P((const char *, char **));
 
 __END_DECLS
